make thumbnail change for all layers visible to admins and move thumbnail into module
